Tests/LibWeb/Text/expected/HTML/image-bitmap-from-video.txt

First error: InvalidStateError: image argument is not usable
Second result: [object ImageBitmap]
